Added HasEnclosure mixin and updated data model.
