+++ itopapi/model/features/__init__.py
@@ -17,3 +17,4 @@ from itopapi.model.features.hasOSVersion import HasOSVersion
 from itopapi.model.features.hasOSLicence import HasOSLicence
 from itopapi.model.features.hasVirtualHost import HasVirtualHost
 from itopapi.model.features.hasRack import HasRack
+from itopapi.model.features.hasEnclosure import HasEnclosure
